requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart>=0.0.9
openai>=1.40.0,<2.0.0
python-dotenv==1.0.1
Pillow>=11.1.0